+++ src/ad/cppgtfs/gtfs/Frequency.h
@@ -17,7 +17,7 @@ namespace gtfs {
 
 class Frequency {
  public:
-  Frequency(const Time& startTime, const Time& endTime, uint16_t headwaySecs,
+  Frequency(const Time& startTime, const Time& endTime, uint32_t headwaySecs,
             bool exactTimes)
       : _startTime(startTime),
         _endTime(endTime),
@@ -26,12 +26,12 @@ class Frequency {
 
   const Time& getStartTime() const { return _startTime; }
   const Time& getEndTime() const { return _endTime; }
-  uint16_t getHeadwaySecs() const { return _headwaySecs; }
+  uint32_t getHeadwaySecs() const { return _headwaySecs; }
   bool hasExactTimes() const { return _exactTimes; }
 
  private:
   Time _startTime, _endTime;
-  uint16_t _headwaySecs;
+  uint32_t _headwaySecs;
   bool _exactTimes;
 };
